requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pyyaml==6.0.2
markdown==3.7
openpyxl==3.1.5
python-docx==1.1.2
